move components into its own project funding folder
